[Android] New c-tor signature for FeedSurfaceCoordinator
Chromium change:
https://chromium-review.googlesource.com/c/chromium/src/+/3246335

feed-modularization: Clean up after FeedHooks migration

Simplify some things now that FeedHooks can be used
to create the ProcessScope.

I wanted to get rid of the initialization at startup
in FeedSurfaceTracker / FeedServiceBridge because timing
of the call can be error prone. The Feed component currently
needs a ProcessScope when it prefetches images, and this
shouldn't be tied to initialization of any UI. To do that,
I used Java reflection.

Bug: 1254437, 1274816